# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JyotishOS API",
    "description": "Deterministic Vedic Astrology Engine with Explainable Classical Rules and Multi-System Consensus",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/chart/calculate": {
      "post": {
        "summary": "Calculate Natal Chart",
        "description": "Calculates full natal chart including Divisional charts and Ashtakavarga.",
        "operationId": "calculate_natal_chart_api_chart_calculate_post",
        "parameters": [
          {
            "name": "ayanamsa",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Lahiri",
              "title": "Ayanamsa"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BirthData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KundaliChart"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chart/shadbala": {
      "post": {
        "summary": "Get Shadbala",
        "description": "Calculates 6-fold Shadbala, Bhava Bala, and Planetary Avasthas.",
        "operationId": "get_shadbala_api_chart_shadbala_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BirthData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chart/milan": {
      "post": {
        "summary": "Run Milan",
        "description": "Calculates 36-Guna Ashtakoota and Manglik Dosha matching.",
        "operationId": "run_milan_api_chart_milan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MilanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/charts/calculate": {
      "post": {
        "tags": [
          "v1-charts"
        ],
        "summary": "Calculate Chart",
        "description": "Calculates full natal chart with Shodashavarga and Ashtakavarga.",
        "operationId": "calculate_chart_api_v1_charts_calculate_post",
        "parameters": [
          {
            "name": "ayanamsa",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Lahiri",
              "title": "Ayanamsa"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BirthData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KundaliChart"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/charts/dashas": {
      "post": {
        "tags": [
          "v1-charts"
        ],
        "summary": "Calculate Multi Dashas",
        "description": "Calculates 4 classical Dasha systems simultaneously for a given target date:\n1. Vimshottari (Maha, Antar, Pratyantar)\n2. Yogini (36-year cycle)\n3. Jaimini Chara Dasha (Sign-based)\n4. Ashtottari Dasha (108-year conditional cycle)",
        "operationId": "calculate_multi_dashas_api_v1_charts_dashas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MultiDashaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/charts/shadbala": {
      "post": {
        "tags": [
          "v1-charts"
        ],
        "summary": "Calculate Shadbala Endpoint",
        "description": "Calculates 6-fold Shadbala strengths, Rupas, and planetary rank order.",
        "operationId": "calculate_shadbala_endpoint_api_v1_charts_shadbala_post",
        "parameters": [
          {
            "name": "ayanamsa",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Lahiri",
              "title": "Ayanamsa"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BirthData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/charts/milan": {
      "post": {
        "tags": [
          "v1-charts"
        ],
        "summary": "Calculate Milan Endpoint",
        "description": "Calculates 36-Guna Ashtakoota Milan compatibility and Manglik Dosha analysis.",
        "operationId": "calculate_milan_endpoint_api_v1_charts_milan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MilanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AshtakavargaResult": {
        "properties": {
          "bav": {
            "additionalProperties": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Bav",
            "description": "7 planets x 12 signs bindu matrix (1-based sign indexing in list: indices 0..11 for Aries..Pisces)"
          },
          "sav": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Sav",
            "description": "12 signs Sarva bindu counts, summing to exactly 337"
          },
          "total_bindus": {
            "type": "integer",
            "title": "Total Bindus",
            "default": 337
          },
          "shodhana": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShodhanaResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "bav",
          "sav"
        ],
        "title": "AshtakavargaResult",
        "description": "Bhinna (BAV) and Sarva (SAV) Ashtakavarga charts with Shodhana."
      },
...
      "BirthData": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": "Jataka"
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "title": "Birth Date"
          },
          "birth_time": {
            "type": "string",
            "format": "time",
            "title": "Birth Time"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "timezone_offset": {
            "type": "number",
            "title": "Timezone Offset",
            "default": 5.5
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City",
            "default": "New Delhi"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence",
            "description": "Exact / Approx / Unknown",
            "default": "Exact"
          }
        },
        "type": "object",
        "required": [
          "birth_date",
          "birth_time",
          "latitude",
          "longitude"
        ],
        "title": "BirthData",
        "description": "Input birth data for a profile."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HouseCusp": {
        "properties": {
          "house_number": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "House Number"
          },
          "sign_id": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Sign Id"
          },
          "sign_name": {
            "type": "string",
            "title": "Sign Name"
          },
          "cusp_longitude": {
            "type": "number",
            "title": "Cusp Longitude"
          },
          "lord": {
            "type": "string",
            "title": "Lord"
          },
          "occupants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Occupants",
            "default": []
          },
          "aspecting_planets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aspecting Planets",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "house_number",
          "sign_id",
          "sign_name",
          "cusp_longitude",
          "lord"
        ],
        "title": "HouseCusp",
        "description": "Cusp information for a bhava."
      },
      "JaiminiResult": {
        "properties": {
          "karakas_7": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Karakas 7"
          },
          "karakas_8": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Karakas 8"
          },
          "karakamsha_sign_id": {
            "type": "integer",
            "title": "Karakamsha Sign Id"
          },
          "karakamsha_sign_name": {
            "type": "string",
            "title": "Karakamsha Sign Name"
          },
          "arudha_padas": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Arudha Padas"
          },
          "arudha_pada_names": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Arudha Pada Names"
          },
          "hora_lagna_sign_name": {
            "type": "string",
            "title": "Hora Lagna Sign Name"
          },
          "ghati_lagna_sign_name": {
            "type": "string",
            "title": "Ghati Lagna Sign Name"
          },
          "sri_lagna_sign_name": {
            "type": "string",
            "title": "Sri Lagna Sign Name"
          },
          "indu_lagna_sign_name": {
            "type": "string",
            "title": "Indu Lagna Sign Name"
          },
          "bhava_lagna_sign_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bhava Lagna Sign Name",
            "default": "Aries"
          },
          "pranapada_lagna_sign_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pranapada Lagna Sign Name",
            "default": "Aries"
          },
          "varnada_lagna_sign_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Varnada Lagna Sign Name",
            "default": "Aries"
          },
          "special_lagnas_detail": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Special Lagnas Detail"
          },
          "arudha_details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arudha Details"
          }
        },
        "type": "object",
        "required": [
          "karakas_7",
          "karakas_8",
          "karakamsha_sign_id",
          "karakamsha_sign_name",
          "arudha_padas",
          "arudha_pada_names",
          "hora_lagna_sign_name",
          "ghati_lagna_sign_name",
          "sri_lagna_sign_name",
          "indu_lagna_sign_name"
        ],
        "title": "JaiminiResult"
      },
      "KundaliChart": {
        "properties": {
          "birth_data": {
            "$ref": "#/components/schemas/BirthData"
          },
          "ayanamsa_name": {
            "type": "string",
            "title": "Ayanamsa Name",
            "default": "Lahiri"
          },
          "ayanamsa_value": {
            "type": "number",
            "title": "Ayanamsa Value"
          },
          "lagna_longitude": {
            "type": "number",
            "title": "Lagna Longitude"
          },
          "lagna_sign_id": {
            "type": "integer",
            "title": "Lagna Sign Id"
          },
          "lagna_sign_name": {
            "type": "string",
            "title": "Lagna Sign Name"
          },
          "lagna_degree": {
            "type": "number",
            "title": "Lagna Degree"
          },
          "planets": {
            "additionalProperties": {
              "$ref": "#/components/schemas/PlanetPosition"
            },
            "type": "object",
            "title": "Planets"
          },
          "houses": {
            "items": {
              "$ref": "#/components/schemas/HouseCusp"
            },
            "type": "array",
            "title": "Houses"
          },
          "panchang": {
            "$ref": "#/components/schemas/PanchangData"
          },
          "vargas": {
            "additionalProperties": {
              "$ref": "#/components/schemas/VargaChart"
            },
            "type": "object",
            "title": "Vargas",
            "default": {}
          },
          "ashtakavarga": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AshtakavargaResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "atmakaraka": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atmakaraka"
          },
          "shadbala": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShadbalaSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "jaimini": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JaiminiResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "upagrahas": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UpagrahaResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "birth_data",
          "ayanamsa_value",
          "lagna_longitude",
          "lagna_sign_id",
          "lagna_sign_name",
          "lagna_degree",
          "planets",
          "houses",
          "panchang"
        ],
        "title": "KundaliChart",
        "description": "Complete computed Vedic Chart."
      },
...
      "MilanRequest": {
        "properties": {
          "groom": {
            "$ref": "#/components/schemas/BirthData"
          },
          "bride": {
            "$ref": "#/components/schemas/BirthData"
          }
        },
        "type": "object",
        "required": [
          "groom",
          "bride"
        ],
        "title": "MilanRequest"
      },
      "MultiDashaRequest": {
        "properties": {
          "birth_data": {
            "$ref": "#/components/schemas/BirthData"
          },
          "target_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date"
          },
          "ayanamsa": {
            "type": "string",
            "title": "Ayanamsa",
            "default": "Lahiri"
          }
        },
        "type": "object",
        "required": [
          "birth_data"
        ],
        "title": "MultiDashaRequest"
      },
      "PanchangData": {
        "properties": {
          "tithi_name": {
            "type": "string",
            "title": "Tithi Name"
          },
          "tithi_number": {
            "type": "integer",
            "title": "Tithi Number"
          },
          "tithi_type": {
            "type": "string",
            "title": "Tithi Type"
          },
          "vara_name": {
            "type": "string",
            "title": "Vara Name"
          },
          "nakshatra_name": {
            "type": "string",
            "title": "Nakshatra Name"
          },
          "yoga_name": {
            "type": "string",
            "title": "Yoga Name"
          },
          "karana_name": {
            "type": "string",
            "title": "Karana Name"
          }
        },
        "type": "object",
        "required": [
          "tithi_name",
          "tithi_number",
          "tithi_type",
          "vara_name",
          "nakshatra_name",
          "yoga_name",
          "karana_name"
        ],
        "title": "PanchangData",
        "description": "Classical Panchang elements at time of event/birth."
      },
...
      "PlanetPosition": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude",
            "description": "Sidereal longitude 0-360 degrees"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude",
            "default": 0.0
          },
          "speed": {
            "type": "number",
            "title": "Speed",
            "default": 0.0
          },
          "is_retrograde": {
            "type": "boolean",
            "title": "Is Retrograde",
            "default": false
          },
          "is_combust": {
            "type": "boolean",
            "title": "Is Combust",
            "default": false
          },
          "sign_id": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Sign Id"
          },
          "sign_name": {
            "type": "string",
            "title": "Sign Name"
          },
          "sign_degree": {
            "type": "number",
            "title": "Sign Degree",
            "description": "Degrees within sign (0-30)"
          },
          "nakshatra_id": {
            "type": "integer",
            "maximum": 27.0,
            "minimum": 1.0,
            "title": "Nakshatra Id"
          },
          "nakshatra_name": {
            "type": "string",
            "title": "Nakshatra Name"
          },
          "nakshatra_pada": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Nakshatra Pada"
          },
          "nakshatra_lord": {
            "type": "string",
            "title": "Nakshatra Lord"
          },
          "house_from_lagna": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "House From Lagna",
            "default": 1
          },
          "house_from_moon": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "House From Moon",
            "default": 1
          },
          "dignity": {
            "type": "string",
            "title": "Dignity",
            "description": "exalted, moolatrikona, own, great_friend, friend, neutral, enemy, great_enemy, debilitated",
            "default": "neutral"
          }
        },
        "type": "object",
        "required": [
          "name",
          "longitude",
          "sign_id",
          "sign_name",
          "sign_degree",
          "nakshatra_id",
          "nakshatra_name",
          "nakshatra_pada",
          "nakshatra_lord"
        ],
        "title": "PlanetPosition",
        "description": "Calculated position and astrological state of a graha."
      },
...
      "ShadbalaSummary": {
        "properties": {
          "planets": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ShadbalaPlanetResult"
            },
            "type": "object",
            "title": "Planets"
          },
          "bhava_bala": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Bhava Bala"
          }
        },
        "type": "object",
        "required": [
          "planets",
          "bhava_bala"
        ],
        "title": "ShadbalaSummary"
      },
...
      "UpagrahaResult": {
        "properties": {
          "gulika_sign_name": {
            "type": "string",
            "title": "Gulika Sign Name"
          },
          "gulika_longitude": {
            "type": "number",
            "title": "Gulika Longitude"
          },
          "mandi_sign_name": {
            "type": "string",
            "title": "Mandi Sign Name"
          },
          "mandi_longitude": {
            "type": "number",
            "title": "Mandi Longitude"
          },
          "dhuma_longitude": {
            "type": "number",
            "title": "Dhuma Longitude"
          },
          "vyatipata_longitude": {
            "type": "number",
            "title": "Vyatipata Longitude"
          },
          "parivesha_longitude": {
            "type": "number",
            "title": "Parivesha Longitude"
          },
          "indrachapa_longitude": {
            "type": "number",
            "title": "Indrachapa Longitude"
          },
          "upaketu_longitude": {
            "type": "number",
            "title": "Upaketu Longitude"
          }
        },
        "type": "object",
        "required": [
          "gulika_sign_name",
          "gulika_longitude",
          "mandi_sign_name",
          "mandi_longitude",
          "dhuma_longitude",
          "vyatipata_longitude",
          "parivesha_longitude",
          "indrachapa_longitude",
          "upaketu_longitude"
        ],
        "title": "UpagrahaResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VargaChart": {
        "properties": {
          "varga_code": {
            "type": "string",
            "title": "Varga Code"
          },
          "varga_name": {
            "type": "string",
            "title": "Varga Name"
          },
          "division": {
            "type": "integer",
            "title": "Division"
          },
          "lagna_sign_id": {
            "type": "integer",
            "title": "Lagna Sign Id"
          },
          "lagna_sign_name": {
            "type": "string",
            "title": "Lagna Sign Name"
          },
          "planets": {
            "additionalProperties": {
              "$ref": "#/components/schemas/VargaPlanetPosition"
            },
            "type": "object",
            "title": "Planets"
          }
        },
        "type": "object",
        "required": [
          "varga_code",
          "varga_name",
          "division",
          "lagna_sign_id",
          "lagna_sign_name",
          "planets"
        ],
        "title": "VargaChart",
        "description": "Divisional Chart (D1 to D60)."
      },
      "VargaPlanetPosition": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sign_id": {
            "type": "integer",
            "title": "Sign Id"
          },
          "sign_name": {
            "type": "string",
            "title": "Sign Name"
          },
          "degree_in_varga": {
            "type": "number",
            "title": "Degree In Varga"
          },
          "house_number": {
            "type": "integer",
            "title": "House Number"
          }
        },
        "type": "object",
        "required": [
          "name",
          "sign_id",
          "sign_name",
          "degree_in_varga",
          "house_number"
        ],
        "title": "VargaPlanetPosition"
      },
...
    }
  }
}